Reduce code and improve performance and memory usage
This change is in the code that compares numbers within strings.
The main performance increase comes from creating less variables
and checking individual characters instead of creating substrings
and comparing those (all of which also reduces memory usage).

This gave a speedup of 1.9 on benchmark #3
This also speeds up benchmark #5 if you modify one of the numbers
such that the difference between the numbers occurs earlier